fix(network): preserve 32-bit game packet sizes
WildStar64's Network_SerialiseResolvedMessage path writes the computed packet byte size as a 32-bit field before writing the 16-bit message id. The previous ServerGamePacket constructor cast HeaderSize + Data.Length through ushort, which could truncate large payload totals before packet write or diagnostics.

Keep Size as HeaderSize plus the uint payload length so managed packet metadata matches the client framing width and avoids silent wraparound.

+++ Source/NexusForever.Network/Packet/ServerGamePacket.cs
@@ -24,7 +24,7 @@ namespace NexusForever.Network.Packet
             }
 
             Opcode = opcode;
-            Size   = (ushort)(HeaderSize + Data.Length);
+            Size   = HeaderSize + (uint)Data.Length;
         }
     }
 }
